trace: only print and read args in a way that users know/care about
Do not print or expect to read in args with additional internal
details that users don't know or care about.  For example, users
don't need to know the internal argument calling convention for
Maxima special forms.

This is also important for an upcoming commit.

No problems with the test suite, share test suite or rtest_translator.