Address #4479:  Handle assigning empty string to inchar
We change how setting `inchar` is done.  Now the value of `inchar` is
always a symbol.  When assigning a string to inchar, we internally
prepend a "$" to it and then intern it as a symbol.

To support this, we need to add a new `:setter-method` option that
takes a symbol of function value.  The value is assigned to the new
`setter-method` property.  `mset` is modified to look for this and
call the method to compute the value to be assigned to the variable.

Update `outchar` to have this setter method too.