In MSET, return the value returned by the setter method.
I misunderstood.  `x:inchar:"a";` is supposed to be set to the value
that is assigned to `inchar`.  In this case, that is `$a`, not the
string "a".

To do this, in `mset`, we must return the value from the setter
method.  We currently do this correctly in `assign-prompts`.

Add a comment in globals.lisp that setter methods must return the
value that is actually assigned to the variable.