Add 'special property and don't remove it in kill1-atom
Add the 'special property to $inchar and friends so that we'll treat
these specially when they are killed.

Also, in kill1-atom, symbols with the 'special property were getting
the property removed.  Don't do that because then the special handling
of symbols (in remvalue) with the special property is deleted.  That
makes `kill(inchar)` infloop with a lisp error.