LispWorks compatibility: call LW:ENVIRONMENT-VARIABLE to get an environment variable.
LW:ENVIRONMENT-VARIABLE exists on both Windows and Linux (other systems
were not investigated). Call LW:ENVIRONMENT-VARIABLE in the two places
where it is needed, and cut out the definition of HARLEQUIN-COMMON-LISP:GETENV
(which just punted to LW:ENVIRONMENT-VARIABLE).

Incidentally put in a definition of INTL::GETENV for ABCL
(copied from MAXIMA-GETENV elsewhere).