Allegro compatibility: ACL on Windows complained about top-level IMPORT.
src/init-cl.lisp contained

    (import 'cl-user::run)

at the top level. Allegro Express 11.0 complained about that, although Allegro Express
for Linux did not, and I do not see in the Common Lisp HyperSpec that IMPORT cannot
be at the top level.

Be that as it may, the import is not needed, because the function RUN is never
referenced as MAXIMA::RUN but only as CL-USER::RUN. So I removed it,
since it was superfluous and caused an error for at least one well-known Lisp.