def-simplifier: Check subs/args for subscripted functions
When defining a simplifier for subscripted functions like psi, no checks were
performed to make sure that the number of subscripts and arguments is correct.
This made it possible to write things like psi[0, 1](x, y) without errors.

def-simplifier now uses the already existing (but not used anywhere) function
subargcheck.