More informative errors for subscripted functions
The functionality of "subargcheck" from simp.lisp was moved to defmfun-check.lisp,
since it was not used anywhere except in def-simplifier. For sake of consistency
with arg-count-check, it has been renamed to sub-arg-count-check, and its
parameters now match the style of arg-count-check. It generates more informative
error messages telling the user exactly what was wrong about subscripts and/or
arguments.