risplit: only use $conjugate when it helps
This fixes bug #4507.

risplit uses $conjugate when it doesn't know how to process an operator, but
$conjugate does. But $conjugate doesn't always give useful results for risplit.
This can result in $rectform answers that don't have an explicit imaginary term,
but are not real-valued, e.g. rectform(gamma(z)) for complex z.

With this change, risplit only uses the result of $conjugate if it doesn't
contain '$conjugate. This also improves results of certain integrals, one of
which was tested in rtestint.mac, so one test had to be changed.

A new test has been added to rtest16.mac.