simp-set: only simplify elements if necessary
simp-set, the simplifier for sets (e.g. {a, b, c}) was using its second argument
incorrectly. The second argument to a simplifier tells it whether the arguments
(in this case: the set elements) are already simplified.

simp-set always called simplifya on the elements and passed the second argument
over to the next level. This is wrong!

The correct behavior: If the second argument is NIL, call (simplifya arg NIL) for
all arguments arg. If the second argument is T, there's no need to call simplifya
on the arguments at all.