Use simplifier to distribute $conjugate over list, =, ...
This fixes bug #4500: "rectform of set".

This change removes the commutes-with-conjugate property from $matrix, mlist,
$set, mequal, mnotequal and %transpose. $conjugate will still distribute over
them, but using the simplifier's distribute_over mechanism, just like most
(all?) other operators do.

Using the commutes-with-conjugate for non-number-valued functions/operators
caused problems with $rectform/risplit.

We should use commutes-with-conjugate only for number-valued functions (for which
conjugation makes sense), not mix it with the concept of distributing over bags.