Print nice function signature for def-simplifier error messages
Error messages from `def-simplifier` for the wrong number of args or
subscripts for subscripted functions print out a nice Maxima form of
the function including the function name and the names of the
subscripts and args, as defined by the `def-simplifier`.

Also, when both the number of subscripts and args are wrong, we print
the two messages on separate lines to make it easier to see and read.