COMPLEX-ATANH: correctly check for Y = 0
COMPLEX-ATANH was using an inefficient and incorrect way to check whether its
argument Y is zero. It compared Y to an invalid bigfloat expression of the form
((BIGFLOAT) 0 0) using LIKE, which checks for special representations and then
calls ALIKE.

((BIGFLOAT) 0 0) is invalid because it's missing the precision value.

This is also very inefficient, since it's sufficient to check whether Y's
mantissa is zero.