ISLINEAR: catch errors while substituting in 0
ISLINEAR determines whether an expression EXPR is linear in a variable VAR1.
It does so by differentiating EXPR w.r.t. VAR1. If the derivative is free of
VAR1, it continues to substitute in 0 for VAR1 in EXPR. However, in some cases,
that can result in errors, for example when EXPR = log(x^2)-2*log(x). The
derivative is zero, but substituting x=0 into EXPR gives a log(0) error.

ISLINEAR now uses NO-ERR-SUB-VAR. If an error occurs, it returns NIL, which
means that the EXPR is not linear in VAR1.

This fixes a bug discovered by Barton Willis.