ALIKE1: Handle X non-cons vs. Y cons quickly
ALIKE1 (X Y) quickly returns NIL when X is a cons, but Y is not.
But it didn't handle the reverse case quickly. For X not cons and Y cons,
it went on to test X for different kind of other types before returning NIL.

With this new test, execution time for this very common case is more than halved.
The performance of ALIKE1 is crucial for the entire system, it's called a lot.