*USE-READABLE-GENSYMS* to control gensym names, 11% speedup
Following a discussion on the mailing list ("Significant performance improvement
of 10-20% "for free", 2025-01-29), this commit introduces the variable
*USE-READABLE-GENSYMS* with the default value :DEBUG. (Using a keyword was
proposed by Raymond Toy.)

This variable controls the use of "readable" gensym names in certain internal
routines related to canonical rational expressions. "Readable" means that the
gensym's name contains a textual representation of the expression that the symbol
stands for, such as SIN(LOG(X)+X^2)2229.

*USE-READABLE-GENSYMS* supports the following values:
- :DEBUG (default) to use readable gensyms only in debug mode, i.e. *MDEBUG* is set.
  This represents a compromise between debuggability for the developers and
  performance for the users (see below).
- T to always use readable gensyms.
- NIL to never use them.

Maxima spends an unreasonably large amount of time generating these "readable"
gensym names, due to the overhead involved in converting expressions to a
textual form. The gensym names are not exposed to the user, and no influence
on the outcome has been observed (and according to Richard Fateman, there
should be none).

Further, according to Richard Fateman, readable gensyms were a "debugging hack".

Without readable gensyms, the test suite execution time is reduced by ~11.3%,
and the memory consumption is reduced by ~8.5% on SBCL 2.5.2.