$MAKELIST: simplify resulting list
Every element added to the list created by $MAKELIST is guaranteed to be
simplified, because it is the result of calling MEVAL, which in turn calls
SIMPLIFYA.

Therefore, call SIMPLIFYA on the resulting list with the second argument set
to T, indicating that the arguments (list elements) are already simplified.

This makes sure that the resulting list is marked as simplified, and the
following call to SIMPLIFYA from within MEVAL won't call SIMPLIFYA on every
list element again. This saves one unnecessary simplifier call per list element.