In elliptic-pi-complete, to return a Maxima complex number
In elliptic-pi-complete, we were converting a Lisp complex number to a
Maxima complex number.  However, all the callers are expecting Lisp
numbers, so this was causing an error.  Thus, remove the call to TO
that did the conversion.

The example, reported by Dimiter Prodanov on the mailing list, Mar 6,
2025, is:

elliptic_pi(2, float(%pi/2), 4)