append, cons, endcons, reverse, sort, sublist: simplifya's 2nd arg. = T
After careful testing and consideration, revise the logic introduced by
commit [ab1ed2]. According to the general convention for Maxima functions,
assume that the arguments are already simplified.

If the arguments are not simplified, it's either a bug or intentionally so,
and it's not the function's responsibility to simplify them.

For example, in "cons(x, L)", $cons can assume that both x and L are
simplified. L being simplified means that all its arguments (e.g. list elements)
are simplified. After consing x to L's arguments, $cons can therefore always
call SIMPLIFYA with the second argument set to T. The newly created expression
still needs to be simplified, and will be, but not its arguments.