WHOLE-INTSUBS: use new limits given by MAKE-DEFINT-ASSUMPTIONS
WHOLE-INTSUBS calls MAKE-DEFINT-ASSUMPTIONS, which can possibly change the limits
of integration. But it ignored these new limits. A comment said that it should
probably never happen that the limits are changed. But for the following integral
from the test suite, it does happen:

integrate(exp(-x^2-1/x^2),x,-inf,inf);

In this case, ignoring the new limits and keeping the old ones doesn't lead to a
wrong result, maybe by luck. In any case, it seems reasonable to always use the
new limits returned by MAKE-DEFINT-ASSUMPTIONS.

This also gets rid of a compiler warning about NEW-LL and NEW-UL being unused.