ZTOINF, MTOINF, SSP, SCMP: add assertions for UL/LL
These functions related to definite integration take LL and/or UL
(lower/upper limit) as an argument, but implicitly assume that they are
$MINF, 0 or $INF.

Add assertions so that wrong usage leads to an error instead of a wrong result.

SSP was called in one place with LL = $MINF when it should have been 0.