Remove unused/broken $ROTATE_PTS and $ROTATE_LIST
A comment in plot.lisp mentioned that $ROTATE_PTS and $ROTATE_LIST aren't used
anywhere and should be removed. They are indeed not used anywhere, and they
are broken, because the definitions of the variables *ROT* and $ROT were already
commented out, and $ROTATE_PTS and $ROTATE_LIST depended on them. Indeed, calling
$ROTATE_PTS gives a Lisp error.

Removing the functions also gets rid of compiler warnings.