KINDP: early exit when symbol has no DATA (1.5% speedup)
KINDP is called extensively during simplification, mostly to check whether a
symbol is declared constant. Such declarations always lead to a DATA entry
being created (or extended) in the symbol's property list.

KINDP can be sped up significantly by checking first if there is a DATA entry
in the first place. If not, return NIL.

This speeds up the entire test suite by 1.5%.