Inline CPLUS, CTIMES, CDIFFERENCE & CMOD, add NORMALIZED-MODULUS
Tests have shown that inlining CPLUS, CTIMES, CDIFFERENCE and CMOD can result
in significant speed-ups for CRE-/polynomial-intensive workloads.

To reduce code bloat caused by inlining, extract the part of CMOD that
"normalizes" the result into a new function NORMALIZED-MODULUS, which is not
inlined. This function is only called when MODULUS is set.

This choice provides the best speed-up for the case when MODULUS is NIL, which
is the expected case.