ALIKE1, SIMPLIFYA, EQTEST: use inlined MEMBER 'SIMP/'ARRAY
Introduce FIND-MEMBER-EQ, an inlined version of MEMBER, and use it in the
time-critical functions ALIKE1, SIMPLIFYA and EQTEST for checking whether an
expression has a 'SIMP or 'ARRAY flag.

FIND-MEMBER-EQ is so small that inlining it instead of calling the regular
MEMBER function introduces almost no code size overhead. But it is significantly
faster because no function call is necessary on the assembly level.

Results with SBCL 2.5.2:
(ALIKE1 '((MPLUS SIMP) A B) '((MPLUS SIMP) B C)): 25% faster
(ALIKE1 '((MPLUS SIMP) A B) '((MPLUS SIMP) A B)): 20% faster