algfac/factor/mat.lisp: don't make B, B2, L, P special
All uses of B, B2, L and P have been checked, they do not need to be special.
There were two instances in mat.lisp where (SETQ L ...) was done without
L appearing as a function argument, in a PROG, LET, etc., but it was used in
a purely lexical way.