Need a space between function name and paren
Add a space between the function name and the opening parenthesis for
the parameters.  This is needed for `ext_diff` and `lie_diff`.  This
is required AFAICT by texinfo.

Adding the space fixes the issue where we had a warning that we had
HTML entries for "lie_diff(vector1,[vector2,form])" and
"ext_diff(form)" that were not in text entries.  However, there are
still two HTML entries not in text entries.  We no longer have a
warning that there are 2 text entries not in the the HTML entries.