Fix #4567: gnuplot_command should be a string
As mentioned in #4567, `gnuplot_command` could be set to anything.
This would produce a lisp error when trying to plot something.

Now, make sure it's set to a string.  If not, a nice Maxima error
message is displayed.  This is done by using `defmvar` with a
:setting-predicate set to the new function `string-predicate` that
checks for a string.

The other defvar's in plot.lisp that clearly take strings are also
modified to use `defvmar` and `:setting-predicate`.