Print nice function signature for def-simplifier error messages
For non-subscripted functions, we print a nice function name,
including args so it's easier to see what the problem is with the
call.

Move all the related functions from muutils.lisp to defmfun-check.lisp
because we need these fairly early in the build.

This change requires adding an extra (optional) arg to arg-count-check
so it can print the pretty name.  This means wna-err also needs an
optional arg for the prett name too.

With this change, we get something like
```
(%i9) atan2(1,2,3);
atan2(y, x): expected exactly 2 arguments but got 3: [1, 2, 3]
 -- an error. To debug this try: debugmode(true);
```

Previously, we would get
```
(%i1) atan2(1,2,3);

atan2: expected exactly 2 arguments but got 3: [1, 2, 3]
```