Fix #4577:  Add scrollbar to examples and background for navbar
Add a background color for the navigation bar to make it stand out
from the actual text.

Add a scrollbar when needed for examples so that the text doesn't
extend past the background color, which is kind of ugly.  This only
happens on narrow screens or if the example text has really long
lines.  For a large screen, the scrollbar does not show up.