Regenerate examples in Integration.texi.m4
Run update_examples.  To make this look right, we had to add assume's
in a few places so we don't ask (new) questions.  We also had to
remove "input:pos" in a few places because we don't ask that question
any more.

Finally, in a few places we had to tweak the output to make it look
right.  These places start with "@c WARNING".