Fix mismatch between html and text indices
For some reason the text help index doesn't include "Package
zeilberger" or "Error and warning messages", but the html index does.

This is a hack to make the html index match the text index.  To do
this, we modify add-entry to check if the TOC entry matches one of the
above.  If so, it's not placed in the html index.