PARSE-TEXINFO-VERSION handles texinfo dev version
Texinfo dev versions produces version strings like "7.2dev".  Update
`parse-texinfo-version` to handle this by handling the "dev" part by
treating it as a patch version of 99.  This assumes there will never
be 99 patches for any texinfo version.

This isn't perfect and could be made a lot more robust, but for now
it's good enough to handle all of the texinfo versions I've run
across.