Add new option :skip-properties to def-simplifier
Instead of letting :custom-defmfun control whether certain properties
are not set, add a new option :skip-properties that is a list of
properties that should not be set.  Thus, :custom-defmfun just means
that the default defmfun is not defined.

Update the simplifiers for realpart and imagpart to use the new option
to skip setting of alias and reversealias.