Remove the custom defmfun for cabs; update the simplifier
By getting rid of the defmfun, we can make things more consistent.  As
Barton showed, the current code says `'cabs(x*y)` returns `abs(x*y)`
instead of `abs(x)*abs(y)`.  This is because the simplifier for cabs
doesn't handle this case.  But `cabs(x*y)` returns the simplified
result because the defmfun handle this.

Updating the simplifier fixes this difference.