Make --list-avail show correct version and be consistent
Currently, "maxima-local --list-avail" would print the version as the
Lisp version of the default lisp.  This is not the Maxima version that
would be expected to be printed.

To fix this, modify maxima-local.in to get the version from the
autoconf variable VERSION, which is the Maxima version.

Next, "maxima-local -g --list-avail" would use Lisp to print out the
available lisps, but the paths used don't work for a local build.
Modify maxima-local.ing to define the envvar "MAXIMA_LOCAL" so that
maxima knows that it's being run from the maxima-local script instead
of maxima.

Modify `list-avail-action` to look for this envvar and search for
pathnames of the form "src/binary-*".  If we find it, print out the
lisp name along with the version, as done by the script in
maxima-local.

This makes "maxima-local -g --list-avail" produce results that are
consistent with just "maxima-local --list-avail".