Update get-dirs to support cmucl and add some comments
`get-dirs` was returning all the files in the given `path` for cmucl
because a trailing slash was not appended.  Append the slash.

Also add some docstrings and comments to `get-dirs`.

Finally, add comments to `unix-like-dirname` because I was confused on
what `unix-like-dirname` was returning.  It is doing what the program
`dirname` does.