Use HERE docs for variables and fix a typo
To make it easier to read and modify the values for HIGHLIGHT1,
HIGHLIGHT2, and CUSTOM_PYGMENTS_CSS use a HERE document.  Then use
`tr` to replace the newlines with spaces.

Finally fix a typo in the filenames for the pygments css files.  Also
set the appropriate media values for the light and dark modes.