Package gentran: create test cases which output C, Fortran, and Ratfor code for a given expression.
In each case, the expected values, as calculated by Maxima, are baked
into the generated code, and the values calculated by the translated
code are compared to those.

Command lines: to generate the translated code,

  $ maxima -b compare_var_translations.mac

where maxima is a recent enough (post-109691d) version.
To compile and run the code for each one,

  $ gcc foomain.c -o foomain-c -lm
  $ ./foomain-c

  $ f77 foomain.f -o foomain-f
  $ ./foomain-f

  $ ratfor foomain.rat > foomain-rat.f
  $ f77 foomain-rat.f -o foomain-rat
  $ ./foomain-rat

with:

  * gcc 9.4.0
  * f77 reporting its version as "GNU Fortran (Ubuntu 5.5.0-12ubuntu1~16.04) 5.5.0 20171010"
  * ratfor built from https://github.com/arnoldrobbins/ratfor as of commit 419cf3