Use CSS variables to set the colors for light and dark mode
Instead of having a completely separate @media section for dark mode,
use variables to define the colors for light and dark modes.

This makes the CSS file smaller and hopefully, it will be easier set
colors for the different modes since all the info is in basically one
place instead of duplicating all the different elements into a
separate @media for dark mode.