Extend "for" loop syntax to iterate over hash tables.
The basic syntax is: for [k, v] in hh do ...
where hh is a undeclared array, "fast" array (i.e., a hash table created
with use_fast_arrays: true), or created by make_array(hashed, ...).

[k, v] is a key/value pair. The pairs are processed in the order
of keys as returned by arrayinfo.

k is always a list, even if the hash table has only one key
and therefore the list has exactly one element.

Destructuring assignment is recognized, e.g.:

  for [[k], v] in hh1 do ...
  for [[k1, k2], v] in hh2 do ...
  for [[k1, k2, k3], v] in hh3 do ...

where hh1, hh2, and hh3 are hash tables which have,
respectively, one, two, and three keys.

Destructuring assignment also applies to the existing loop syntax
for iterating over lists, e.g.:

  for [x, y] in [[11, 12], [13, 9], [19, 29]] do ...

This is a freebie -- it would be more work to apply destructuring
assignment only to hash tables, so I let it stand.

There is one bug of which I'm aware: when destructuring assignment is
applied and one of the loop variables is restricted by an ASSIGN helper
to non-symbol values (e.g., linel, which has an ASSIGN helper to
restrict assigned values to integers), *and* the variable is unbound
in the Lisp sense (i.e., BOUNDP returns NIL) before starting the
iteration, then there is an error at the end of the iteration.

From what I can tell, there are no such variables in Maxima, and it's
not easy to create one. Such a variable can be created by Lisp
programming (to create an assignment helper and attach it via the
ASSIGN property), or the following Maxima incantation:

  define_variable (foo, 1111, integer);
  remvalue (foo);

foo has an ASSIGN property (created by define_variable) and it is
unbound by remvalue. Now the following loop triggers the bug:

  for [foo] in [[11], [22], [33]] do print (foo);

From reading the reference manual, it appears that mode_declare plus
mode_checkp, and followed by remvalue, might also work, although it
appears mode_checkp has no effect, from what I can tell.

I will devote time to fixing the bug, but it seems like a not too
serious problem, given the obscurity of the way to trigger it.